src_bot/requirements.txt
python-telegram-bot==20.3
runpod
qdrant-client
transformers
scikit-learn
langchain